+++ stag_lib/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the library are documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- [Issue #56](https://github.com/staglibrary/stag/issues/56): Increase convergence speed of eigenvalue calculation
+
 ## [0.4.0] - 2023-01-16
 ### Fixed
 - [Issue #45](https://github.com/staglibrary/stag/issues/45): Make the LocalGraph class pure virtual